chore: move dashboard helper to helpers directory
