+++ pages/dashboard/mentor.tsx
@@ -12,7 +12,7 @@ import ProjectTable from "@/components/tables/ProjectTable";
 import useFetch, { isFetching } from "@/hooks/useFetch";
 import useAuth from "@/contexts/useAuth";
 // Helpers
-import { transformTabNameIntoId } from "./Dashboard.helpers";
+import { transformTabNameIntoId } from "@/helpers/dashboard";
 // Type
 import type { NextPage } from "next";
 import { ROLES } from "@/types/roles";
